+++ lib/lotus/sources/postgres.ex
@@ -4,6 +4,7 @@ defmodule Lotus.Sources.Postgres do
   @behaviour Lotus.Source
 
   alias Lotus.Sources.Default
+  alias Lotus.SQL.FilterInjector
 
   @postgrex_error Module.concat([:Postgrex, :Error])
 
@@ -182,6 +183,37 @@ defmodule Lotus.Sources.Postgres do
   end
 
   @impl true
+  def explain_plan(repo, sql, params, opts) do
+    explain_sql = "EXPLAIN (FORMAT JSON) " <> sql
+    search_path = Keyword.get(opts, :search_path)
+
+    result =
+      repo.transaction(fn ->
+        repo.query!("SET LOCAL transaction_read_only = on")
+        if search_path, do: repo.query!("SET LOCAL search_path = #{search_path}")
+        repo.query(explain_sql, params)
+      end)
+      |> case do
+        {:ok, query_result} -> query_result
+        {:error, err} -> {:error, err}
+      end
+
+    case result do
+      {:ok, %{rows: [[json]]}} ->
+        plan_text =
+          case json do
+            binary when is_binary(binary) -> binary
+            data -> Lotus.JSON.encode!(data)
+          end
+
+        {:ok, plan_text}
+
+      {:error, err} ->
+        {:error, format_error(err)}
+    end
+  end
+
+  @impl true
   def resolve_table_schema(repo, table, schemas) do
     sql = """
     SELECT table_schema
@@ -195,6 +227,17 @@ defmodule Lotus.Sources.Postgres do
       {:ok, %{rows: [[schema]]}} -> schema
       _ -> nil
     end
+  end
+
+  @impl true
+  def quote_identifier(identifier) do
+    escaped = String.replace(identifier, "\"", "\"\"")
+    ~s("#{escaped}")
+  end
+
+  @impl true
+  def apply_filters(sql, filters) do
+    FilterInjector.apply(sql, filters, &quote_identifier/1)
   end
 
   defp format_postgres_type("character varying", char_len, _, _) when not is_nil(char_len),